+++ scilab/modules/graphics/includes/Champ.h
@@ -23,10 +23,7 @@
 #include "machine.h" /* C2F */
 
 GRAPHICS_IMPEXP int C2F(champ) (double *x, double *y, double *fx, double *fy, int *n1, int *n2, char *strflag, double *brect, double *arfact, int lstr) ;
-GRAPHICS_IMPEXP int C2F(champ1) (double *x, double *y, double *fx, double *fy, int *n1, int *n2, char *strflag, double *brect, double *arfact, int lstr) ;
-
 GRAPHICS_IMPEXP void champg(char *name, int colored, double *x, double *y, double *fx, double *fy, int *n1,
                             int *n2, char *strflag, double *brect, double *arfact, int lstr);
-
 
 #endif