+++ scilab/modules/graphics/src/c/getHandleProperty/get_datatip_property.c
@@ -44,27 +44,6 @@ void* get_tip_orientation_property(void* _pvCtx, int iObjUID)
 }
 
 /**
- * Old z_componet property, warns the user
- */
-void* get_tip_z_component_property(void* _pvCtx, int iObjUID)
-{
-    char * tip_display_components;
-    getGraphicObjectProperty(iObjUID, __GO_DATATIP_DISPLAY_COMPONENTS__, jni_string, (void **)&tip_display_components);
-
-    //Only warns if the property exists for the object.
-    if (tip_display_components == NULL)
-    {
-        Scierror(999, _("'%s' property does not exist for this handle.\n"), "z_component");
-    }
-    else
-    {
-        Scierror(999, _("'%s' property is obsolete and will be removed, use '%s' instead.\n"), "z_component", "display_components");
-    }
-
-    return NULL;
-}
-
-/**
  * Get display mode for datatips
  */
 void* get_datatip_display_mode_property(void* _pvCtx, int iObjUID)