+++ scilab/modules/gui/includes/gw_gui.h
@@ -68,7 +68,6 @@ GUI_IMPEXP int sci_datatip_set_orient(char *fname, void* pvApiCtx);
 
 // Scilab 6
 GUI_IMPEXP int sci_about(char* fname, void* pvApiCtx);
-GUI_IMPEXP int sci_mpopup(char* fname, void* pvApiCtx);
 GUI_IMPEXP int sci_fire_closing_finished (char* fname, void* pvApiCtx);
 
 GUI_IMPEXP int sci_figure(char *fname, void* pvApiCtx);