+++ scilab/modules/history_manager/sci_gateway/cpp/history_manager_gw.cpp
@@ -30,7 +30,6 @@ int HistoryManagerModule::Load()
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"historysize",   &sci_historysize, MODULE_NAME));
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"removelinehistory",   &sci_removelinehistory, MODULE_NAME));
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"resethistory", &sci_resethistory, MODULE_NAME));
-    symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"saveafterncommands",   &sci_saveafterncommands, MODULE_NAME));
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"saveconsecutivecommands",   &sci_saveconsecutivecommands, MODULE_NAME));
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"savehistory", &sci_savehistory, MODULE_NAME));
     symbol::Context::getInstance()->addFunction(types::Function::createFunction(L"sethistoryfile",   &sci_sethistoryfile, MODULE_NAME));