+++ scilab/modules/slint/src/cpp/DeprecatedChecker.cpp
@@ -92,15 +92,26 @@ std::unordered_map<std::wstring, std::wstring> DeprecatedChecker::initDep()
     // TODO: get this list from a conf file
     std::unordered_map<std::wstring, std::wstring> map;
 
-    // Scilab 6.1.0 => 6.1.x
+    // Scilab 6.1.x => 2023.0.0
     map.emplace(L"fplot2d", L"plot");
     map.emplace(L"xdel", L"close");
     map.emplace(L"xname", L"gcf().figure_name");
-
-    // Scilab 6.0.x => 6.1.0
+    map.emplace(L"soundsec", L"0 : 1/freq : t*(1-%eps)");
+    map.emplace(L"importgui", L"uiSpreadsheet");
+    map.emplace(L"closeEditvar", L"close editvar");
+    
+    // Scilab 6.1.0 => 6.1.x
+    map.emplace(L"scatter3", L"scatter3d");
+    map.emplace(L"get_figure_handle", L"findobj(\'figure_id\',n)");
+    map.emplace(L"noisegen", L"grand");
     map.emplace(L"%sn", L"ellipj");
     map.emplace(L"champ1", L"champ.colored");
-    map.emplace(L"closeEditvar", L"close editvar");
+    map.emplace(L"saveafterncommands", L"");
+    map.emplace(L"setPreferencesValue", L"xmlSetValues");
+    map.emplace(L"sysdiag", L"blockdiag");
+    map.emplace(L"ric_desc", L"riccati");
+    
+    // Scilab 6.0.x => 6.1.0
     map.emplace(L"dirname", L"fileparts");
     map.emplace(L"_d", L"_");
     map.emplace(L"dgettext", L"gettext");
@@ -108,20 +119,13 @@ std::unordered_map<std::wstring, std::wstring> DeprecatedChecker::initDep()
     map.emplace(L"denom", L".den");
     map.emplace(L"eval", L"evstr");
     map.emplace(L"frexp", L"log2");
-    map.emplace(L"get_figure_handle", L"findobj(\'figure_id\',n)");
     map.emplace(L"getPreferencesValue", L"xmlGetValues");
     map.emplace(L"hypermat", L"zeros|matrix");
     map.emplace(L"lstsize", L"size");
     map.emplace(L"nanmin", L"min");
     map.emplace(L"nanmax", L"max");
-    map.emplace(L"noisegen", L"grand");
     map.emplace(L"numer", L".num");
-    map.emplace(L"ric_desc", L"riccati");
-    map.emplace(L"saveafterncommands", L"");
-    map.emplace(L"scatter3", L"scatter3d");
-    map.emplace(L"setPreferencesValue", L"xmlSetValues");
     map.emplace(L"square", L"replot");
-    map.emplace(L"sysdiag", L"blockdiag");
     map.emplace(L"with_tk", L"with_module('tclsci')");
     map.emplace(L"xgetech", L"gca");
     map.emplace(L"xinfo", L"gcf().info_message");
@@ -247,6 +251,7 @@ std::unordered_map<std::wstring, std::wstring> DeprecatedChecker::initDep()
     map.emplace(L"with_texmacs", L"");
     map.emplace(L"xbasr", L"");
     map.emplace(L"xselect", L"show_window");
+    map.emplace(L"mpopup", L"uicontextmenu");
 
     // Scilab 5.3.3 => 5.4.0
     map.emplace(L"MSDOS", L"getos");