+++ scilab/modules/ui_data/sci_gateway/cpp/sci_editvar.cpp
@@ -228,18 +228,6 @@ int sci_editvar(char * fname, void* pvApiCtx)
         return 0;
     }
 
-    /* Closing the editor
-       ------------------
-    */
-    if (strcmp(pStVarOne, "close") == 0)
-    {
-        EditVar::closeVariableEditor(getScilabJavaVM());
-        freeAllocatedSingleString(pStVarOne);
-        return 0;
-    }
-    /* Otherwise : go on editing */
-
-
     /* get address of the variable*/
     sciErr = getVarAddressFromName(pvApiCtx, pStVarOne, &piAddr);
     if (sciErr.iErr)